+++ editor/src/editor/panels/editor_panel_fonts.cpp
@@ -60,7 +60,7 @@ void EditorPanelFonts::DrawContents() {
 
     ImGui::PushID(this);
     if (ImGui::BeginListBox("##font_list", ImVec2(-FLT_MIN, 0))) {
-        for (int i = 0; i < static_cast<int>(fontNames.size()); ++i) {
+        for (size_t i = 0; i < fontNames.size(); ++i) {
             auto const& entryInfo = fontNames[i];
             bool const selected = m_selectedIndex == i;
             if (ImGui::Selectable(entryInfo.c_str(), selected)) {
